Fixed Sql error

+++ Application/Modules/NFTsModule/Commands/NFTAddCommand/NFTAddRequest.cs
@@ -8,8 +8,8 @@ namespace Application.Modules.NFTsModule.Commands.NFTAddCommand
     {
         public string Title { get; set; }
         public string Description{ get; set; }
-        public short Price { get; set; }
-        public short HighestBid { get; set; }
+        public double Price { get; set; }
+        public double HighestBid { get; set; }
         public IFormFile Image { get; set; }
         public int CreatorId { get; set; }
     }
