+++ Application/Modules/NFTsModule/Queries/NFTGetByIdQuery/NFTGetByIdRequestDto.cs
@@ -12,8 +12,8 @@ namespace Application.Modules.NFTsModule.Queries.NFTGetByIdQuery
         public int CreatorId { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public short Price { get; set; }
-        public short HighestBid { get; set; }
+        public double Price { get; set; }
+        public double HighestBid { get; set; }
         public string ImagePath { get; set; }
     }
 }
