+++ DataAccessLayer/Configurations/NFTEntityTypeConfiguration.cs
@@ -12,8 +12,8 @@ namespace DataAccessLayer.Configurations
             builder.Property(m => m.CreatorId).HasColumnType("int");
             builder.Property(m => m.Title).HasColumnType("nvarchar").HasMaxLength(400).IsRequired();
             builder.Property(m => m.Description).HasColumnType("nvarchar").HasMaxLength(400).IsRequired();
-            builder.Property(m => m.Price).HasColumnType("smallint").IsRequired();
-            builder.Property(m => m.HighestBid).HasColumnType("smallint").IsRequired();
+            builder.Property(m => m.Price).HasColumnType("float").IsRequired();
+            builder.Property(m => m.HighestBid).HasColumnType("float").IsRequired();
             builder.Property(m => m.ImagePath).HasColumnType("varchar").HasMaxLength(100).IsRequired();
 
             builder.ConfigureAuditable();
