+++ Domain/Models/Entities/NFT.cs
@@ -13,8 +13,8 @@ namespace Domain.Models.Entities
         public int CreatorId { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public short Price { get; set; }
-        public short HighestBid { get; set; }
+        public double Price { get; set; }
+        public double HighestBid { get; set; }
         public string ImagePath { get; set; }
     }
 }
